nightly other classes model added
