+++ tests/nightly_pipeline/audio_embedding_models/test_generate.py
@@ -16,7 +16,7 @@ from transformers import AutoProcessor
 
 from QEfficient import QEFFAutoModelForCTC
 
-from ..nightly_utils import get_onnx_and_qpc_size, human_readable
+from ..nightly_utils import NIGHTLY_SKIPPED_MODELS, get_onnx_and_qpc_size
 
 model_config_path = os.path.join(os.path.dirname(__file__), "../configs/validated_models.json")
 with open(model_config_path, "r") as f:
@@ -28,9 +28,11 @@ test_models = config["audio_embedding_models"]
 @pytest.mark.parametrize("model_name", test_models[:1])
 def test_generate_audio_embedding_model(model_name, get_model_config, audio_embedding_model_artifacts):
 
+    if model_name in NIGHTLY_SKIPPED_MODELS:
+        pytest.skip(f"Skipping {model_name} as it is in nightly skipped models list.")
+
     config, pipeline_configs = get_model_config
     compile_params = pipeline_configs["audio_embedding_model_configs"][0].get("compile_params", {})
-    # generate_params = pipeline_configs["audio_embedding_model_configs"][0].get("generate_params", {})
 
     # Retrieve onnx_path from previous stage
     if (
@@ -61,7 +63,6 @@ def test_generate_audio_embedding_model(model_name, get_model_config, audio_embe
 
     onnx_and_qpc_dir = os.path.dirname(onnx_path)
     size = get_onnx_and_qpc_size(onnx_and_qpc_dir)
-    size = human_readable(size)
     # Store all metrics and execution info
     audio_embedding_model_artifacts[model_name].update(
         {
