+++ tests/nightly_pipeline/audio_models/test_export_compile.py
@@ -6,6 +6,8 @@ import pytest
 
 from QEfficient import QEFFAutoModelForSpeechSeq2Seq
 
+from ..nightly_utils import NIGHTLY_SKIPPED_MODELS
+
 model_config_path = os.path.join(os.path.dirname(__file__), "../configs/validated_models.json")
 with open(model_config_path, "r") as f:
     config = json.load(f)
@@ -16,6 +18,9 @@ test_models = config["audio_models"]
 @pytest.mark.parametrize("model_name", test_models[:1])
 def test_export_compile_audio_model(model_name, get_model_config, audio_model_artifacts):
     """Test export and compile of audio model."""
+
+    if model_name in NIGHTLY_SKIPPED_MODELS:
+        pytest.skip(f"Skipping {model_name} as it is in nightly skipped models list.")
 
     config, pipeline_configs = get_model_config
     export_params = pipeline_configs["audio_model_configs"][0].get("export_params", {})
