+++ tests/nightly_pipeline/configs/validated_models.json
@@ -1,24 +1,27 @@
 {
-    "causal_lm_models":[
+    "causal_lm_models": [
         "openai-community/gpt2",
+        "allenai/OLMo-2-0425-1B",
         "tiiuae/falcon-40b",
-        "google/codegemma-2b",  
-        "google/codegemma-7b",  
-        "google/gemma-2b",  
-        "google/gemma-7b",  
-        "google/gemma-2-2b", 
-        "google/gemma-2-9b", 
-        "google/gemma-2-27b", 
-        "bigcode/starcoder",  
-        "bigcode/starcoder2-15b",  
+        "Qwen/Qwen3-30B-A3B-Instruct-2507",
+        "google/codegemma-2b",
+        "google/codegemma-7b",
+        "google/gemma-2b",
+        "google/gemma-7b",
+        "google/gemma-2-2b",
+        "google/gemma-2-9b",
+        "google/gemma-2-27b",
+        "openai/gpt-oss-20b",
+        "bigcode/starcoder",
+        "bigcode/starcoder2-15b",
         "EleutherAI/gpt-j-6b",
-        "ibm-granite/granite-3.1-8b-instruct",  
-        "ibm-granite/granite-guardian-3.1-8b",  
-        "ibm-granite/granite-20b-code-base-8k",  
-        "ibm-granite/granite-20b-code-instruct-8k",  
-        "codellama/CodeLlama-7b-hf", 
-        "codellama/CodeLlama-13b-hf",  
-        "codellama/CodeLlama-34b-hf",  
+        "ibm-granite/granite-3.1-8b-instruct",
+        "ibm-granite/granite-guardian-3.1-8b",
+        "ibm-granite/granite-20b-code-base-8k",
+        "ibm-granite/granite-20b-code-instruct-8k",
+        "codellama/CodeLlama-7b-hf",
+        "codellama/CodeLlama-13b-hf",
+        "codellama/CodeLlama-34b-hf",
         "deepseek-ai/DeepSeek-R1-Distill-Llama-70B",
         "inceptionai/jais-adapted-7b",
         "inceptionai/jais-adapted-13b-chat",
@@ -38,12 +41,14 @@
         "lmsys/vicuna-13b-v1.5",
         "mistralai/Mistral-7B-Instruct-v0.1",
         "mistralai/Codestral-22B-v0.1",
-        "mistralai/Mixtral-8x7B-Instruct-v0.1",
         "mistralai/Mixtral-8x7B-v0.1",
         "microsoft/Phi-3-mini-4k-instruct",
-        "Qwen/Qwen2-1.5B-Instruct"
+        "deepseek-ai/DeepSeek-R1-Distill-Qwen-32B",
+        "Qwen/Qwen2-1.5B-Instruct",
+        "Snowflake/Llama-3.1-SwiftKV-8B-Instruct",
+        "hpcai-tech/grok-1"
     ],
-    
+
     "image_text_to_text_models": [
         "llava-hf/llava-1.5-7b-hf",
         "meta-llama/Llama-3.2-11B-Vision-Instruct",
@@ -52,7 +57,10 @@
         "Llama-4-Scout-17B-16E-Instruct",
         "google/gemma-3-4b-it",
         "Qwen/Qwen2.5-VL-3B-Instruct",
-        "mistralai/Mistral-Small-3.1-24B-Instruct-2503"
+        "mistralai/Mistral-Small-3.1-24B-Instruct-2503",
+        "allenai/Molmo-7B-D-0924",
+        "OpenGVLab/InternVL2_5-1B",
+        "OpenGVLab/InternVL3_5-1B"
     ],
 
     "embedding_models": [
@@ -86,5 +94,4 @@
     "sequence_models": [
         "meta-llama/Llama-Prompt-Guard-2-22M"
     ]
-
 }
