+++ tests/nightly_pipeline/embedding_models/test_export_compile.py
@@ -15,6 +15,8 @@ import torch
 
 from QEfficient import QEFFAutoModel as AutoModel
 
+from ..nightly_utils import NIGHTLY_SKIPPED_MODELS
+
 
 def max_pooling(last_hidden_states: torch.Tensor, attention_mask: torch.Tensor) -> torch.Tensor:
     """Apply max pooling to the last hidden states."""
@@ -34,6 +36,9 @@ test_models = config["embedding_models"]
 @pytest.mark.parametrize("pooling", [None])
 def test_export_compile_embedding_model(model_name, pooling, get_model_config, embedding_model_artifacts):
     """Test export and compile for embedding models."""
+    if model_name in NIGHTLY_SKIPPED_MODELS:
+        pytest.skip(f"Skipping {model_name} as it is in nightly skipped models list.")
+
     config, pipeline_configs = get_model_config
     export_params = pipeline_configs["embedding_model_configs"][0].get("export_params", {})
     compile_params = pipeline_configs["embedding_model_configs"][0].get("compile_params", {})
