+++ tests/nightly_pipeline/image_text_to_text_models/test_export_compile.py
@@ -7,6 +7,8 @@ import pytest
 from QEfficient import QEFFAutoModelForCausalLM, QEFFAutoModelForImageTextToText
 from QEfficient.utils.test_utils import ModelConfig
 
+from ..nightly_utils import NIGHTLY_SKIPPED_MODELS
+
 model_config_path = os.path.join(os.path.dirname(__file__), "../configs/validated_models.json")
 with open(model_config_path, "r") as f:
     config = json.load(f)
@@ -19,6 +21,9 @@ test_models = config["image_text_to_text_models"]
 def test_export_compile_image_text_to_text_model(
     model_name, kv_offload, image_text_to_text_model_artifacts, get_model_config
 ):
+
+    if model_name in NIGHTLY_SKIPPED_MODELS:
+        pytest.skip(f"Skipping {model_name} as it is in nightly skipped models list.")
 
     config, pipeline_configs = get_model_config
     export_params = pipeline_configs["image_text_to_text_model_configs"][0].get("export_params", {})
@@ -43,7 +48,6 @@ def test_export_compile_image_text_to_text_model(
         )
     export_loading_time = time.time() - export_load_start
 
-    print(qeff_model.model.config)
     # Export time
     print(f"\nExporting for model: {model_name}")
     export_start = time.time()
