+++ tests/nightly_pipeline/image_text_to_text_models/test_generate.py
@@ -25,7 +25,7 @@ from transformers import (
 from QEfficient import QEFFAutoModelForCausalLM, QEFFAutoModelForImageTextToText
 from QEfficient.utils.test_utils import InternProcessor, ModelConfig
 
-from ..nightly_utils import get_onnx_and_qpc_size
+from ..nightly_utils import NIGHTLY_SKIPPED_MODELS, get_onnx_and_qpc_size
 
 model_config_path = os.path.join(os.path.dirname(__file__), "../configs/validated_models.json")
 with open(model_config_path, "r") as f:
@@ -39,6 +39,9 @@ test_models = config["image_text_to_text_models"]
 def test_generate_image_text_to_text_model(
     model_name, kv_offload, image_text_to_text_model_artifacts, get_model_config
 ):
+
+    if model_name in NIGHTLY_SKIPPED_MODELS:
+        pytest.skip(f"Skipping {model_name} as it is in nightly skipped models list.")
 
     config, pipeline_configs = get_model_config
     compile_params = pipeline_configs["image_text_to_text_model_configs"][0].get("compile_params", {})
