+++ tests/nightly_pipeline/sequence_models/test_export_compile.py
@@ -14,6 +14,8 @@ import pytest
 
 from QEfficient import QEFFAutoModelForSequenceClassification
 
+from ..nightly_utils import NIGHTLY_SKIPPED_MODELS
+
 model_config_path = os.path.join(os.path.dirname(__file__), "../configs/validated_models.json")
 with open(model_config_path, "r") as f:
     config = json.load(f)
@@ -24,6 +26,10 @@ test_models = config["sequence_models"]
 @pytest.mark.parametrize("model_name", test_models)
 def test_export_compile_sequence_model(model_name, get_model_config, sequence_model_artifacts):
     """Test export and compile for sequnce models."""
+
+    if model_name in NIGHTLY_SKIPPED_MODELS:
+        pytest.skip(f"Skipping {model_name} as it is in nightly skipped models list.")
+
     config, pipeline_configs = get_model_config
     export_params = pipeline_configs["sequence_model_configs"][0].get("export_params", {})
     compile_params = pipeline_configs["sequence_model_configs"][0].get("compile_params", {})
